Fix TypeError in updateEventReadout: cast float progress to int for Qt5
QProgressBar.setValue() requires int in Qt5/PySide6, but the callback
from qreduce.py passes float values (0., 0.1, etc.). This caused a
crash when opening any file with a progress callback (e.g. run 25899).

+++ quicknxs/main_gui.py
@@ -2200,7 +2200,7 @@ class MainGUI(QtWidgets.QMainWindow):
     When reading event mode data this is the callback
     used after each finished channel to indicate the progress.
     '''
-    self.eventProgress.setValue(progress*100)
+    self.eventProgress.setValue(int(progress*100))
     # make sure the update is shown in the interface
     self.eventProgress.update()
